backend/requirements.txt
fastapi
uvicorn
python-binance
pandas
sqlalchemy
python-dotenv
pydantic
websockets
